Add Features and Walthrough.pdf
AutoPublisher now fixes broken data source links.
A "Retry Last" button has been added to the last ("One" or "All")
window.  It will appear when one or maps are in the retry list, which is
created at the end of each execution.  If maps cannot be published they
are added to the retry list.  The user can fix the issues and easily
make a second attempt at publishing using the "Retry Last" button.